Bazel: Use toolchains rather than select for bzlmod compatibility
By using toolchains and exec_compatbile_with we can use the correct
toolchain for the execution platform.

select was not operating on the exec platform, which meant that files
weren't getting to the right places with bzlmod + Bazel 8.

Toolchains are set up for all platforms that were previously supported.
If you attempt to use a version of Emscripten that doesn't have an arm64
Linux build to build wasm on that platform, then it will error as it did
previously.